Move find_duplicates.
